Refactor imports for consistency: move _CommandsTemplate and _CommandFactory to utils; update import paths
