+++ CityOfBinds/src/content_managers/utils/commands_template.py
@@ -1,7 +1,7 @@
 from .command_factory import _CommandFactory
-from ....game.utils import _Power
-from ....game.utils import _CommandGroup, CommandGroupConstants
-from .....utils import ListTemplate, Pool
+from ...game.utils import _Power
+from ...game.utils import _CommandGroup, CommandGroupConstants
+from ....utils import ListTemplate, Pool
 
 
 class _CommandsTemplate(_CommandGroup, ListTemplate):
